refactor: login routes to reflect simpler module structure
