+++ web-client/src/modules/login/index.ts
@@ -1,9 +1,9 @@
 import { Module } from 'src/types/module';
 
-import ContentPage from './pages/ContentPage';
+import Routes from './routes';
 
 const loginModule: Module = {
   path: '/login',
-  component: ContentPage,
+  component: Routes,
 };
 export default loginModule;
