--- web-client/src/modules/login/pages/routes/LoginRoute/LoginRoute.tsx
@@ -1,22 +1,0 @@
-import get from 'lodash/get';
-import React, { ReactElement } from 'react';
-import { useLocation } from 'react-router-dom';
-
-import CenteredCard from '../../../../../components/CenteredCard/CenteredCard';
-import GradientBackground from '../../../../../components/GradientBackground/GradientBackground';
-import IntroWrapper from '../../../../../components/IntroComponent/IntroComponent';
-import LoginContainer from '../../../containers/LoginContainer/LoginContainer';
-
-const LoginRoute: React.FC = (): ReactElement => {
-  const location = useLocation();
-  return (
-    <GradientBackground>
-      <CenteredCard>
-        <IntroWrapper>
-          <LoginContainer redirectBack={get(location, 'state.redirectBack')} />
-        </IntroWrapper>
-      </CenteredCard>
-    </GradientBackground>
-  );
-};
-export default LoginRoute;
